refactor: Add support for reportDataInDays and isReportEnabled in StatusPageSubscriberService
