+++ CommonProject/.gitignore
@@ -1,0 +1,29 @@
+# See https://help.github.com/ignore-files/ for more about ignoring files.
+
+# dependencies
+#/backend/node_modules
+/kubernetes
+/node_modules
+.idea
+
+# misc
+.DS_Store
+
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+yarn.lock
+
+**/*/paymentService.test.js
+apiTest.rest
+
+application_security_dir
+container_security_dir
+
+# coverage
+/coverage
+/.nyc_output
+
+/greenlock.d/config.json
+/greenlock.d/config.json.bak
