+++ CommonProject/README.md
@@ -1,0 +1,3 @@
+# OneUptime Common UI Library
+
+The OneUptime Common UI Library is a collection of shared components, utilities that are used across the OneUptime platform. It is designed to be easy to install and use, and to be extensible. This library is built with React and TypeScript. It includes components like buttons, modals, and form elements. It also includes utilities like date formatting, and string manipulation.
