+++ CommonProject/Utils/Uptime/MonitorEvent.ts
@@ -1,0 +1,6 @@
+import ObjectID from "Common/Types/ObjectID";
+import Event from "./Event";
+
+export default interface MonitorEvent extends Event {
+    monitorId: ObjectID;
+}
