+++ CommonProject/jest.config.json
@@ -1,0 +1,44 @@
+{
+   
+ "preset": "ts-jest",
+"testPathIgnorePatterns": [
+        "node_modules",
+        "dist"
+    ],
+    "verbose": true,
+    "globals": {
+        "ts-jest": {
+            "tsconfig": "tsconfig.json",
+            "babelConfig": false
+        }
+    },
+    "moduleFileExtensions": [
+        "ts",
+        "js",
+        "json"
+    ],
+    "transform": {
+        ".(ts|tsx)": "ts-jest"
+    },
+    "testEnvironment": "node",
+    "collectCoverage": false,
+    "coverageReporters": [
+        "text",
+        "lcov"
+    ],
+    "testRegex": "./Tests/(.*).test.ts",
+    "collectCoverageFrom": [
+        "./**/*.(tsx||ts)"
+    ],
+    "moduleNameMapper": {
+        "axios": "axios/dist/node/axios.cjs"
+    },
+    "coverageThreshold": {
+        "global": {
+            "lines": 0,
+            "functions": 0,
+            "branches": 0,
+            "statements": 0
+        }
+    }
+}
