+++ CommonProject/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "@oneuptime/common-project",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "test": "jest --detectOpenHandles",
+        "coverage": "jest --detectOpenHandles --coverage",
+        "compile": "tsc",
+        "clear-modules": "rm -rf node_modules && rm package-lock.json && npm install",
+        "dep-check": "npm install -g depcheck && depcheck ./ --skip-missing=true"
+    },
+    "author": "OneUptime <hello@oneuptime.com> (https://oneuptime.com/)",
+    "license": "Apache-2.0",
+    "devDependencies": {
+        "@faker-js/faker": "^8.0.2",
+        "@types/jest": "^27.5.2",
+        "@types/node": "^17.0.22",
+        "jest": "^27.5.1",
+        "ts-jest": "^27.1.4"
+    },
+    "dependencies": {
+       "Model": "file:../Model",
+       "Common": "file:../Common"
+    }
+}
