+++ CommonUI/package.json
@@ -31,6 +31,7 @@
         "@types/react-highlight": "^0.12.8",
         "@types/react-syntax-highlighter": "^15.5.13",
         "Common": "file:../Common",
+        "CommonProject": "file:../CommonProject",
         "formik": "^2.4.6",
         "history": "^5.3.0",
         "jwt-decode": "^4.0.0",
